Code refactoring.  Simplified Listener and effects
Initializing the Network Listener has been broken into 2 methods.
The first method, that has no args, will get the default args and
pass those valuse to the second method.  This second method, that
has 2 args, will use the provided values to set the listener.

The code that creates the error effect in the Network settings UI
has been moved to the Kaiwa object.  This will allow for other UI
components to use the same error effect for a consistant look and
feel.